Updated passport dependency to >=
Projects using passport 0.3.x will have passport 0.1.15 loaded as a submodule of this module. This will in turn cause some compatibility issues when serializing user objects. When a backwards incompatible version of passport comes around, the dependency should be updated. I have not really tested this for passport other than 0.3.0, but if it worked with 0.1.15 and it works with 0.3.0, I'm assuming it works for everything in between and less than 0.4.0